feat: 메인 헤더 수정
